Revert "Merge branch 'master' of https://github.com/sobotz/deepspace"
This reverts commit 52027bcaf94a58a7d0c88d30d2779a319e81a2c3, reversing
changes made to 848bd18b0d4b80e97d7e284ee7d20d8ab354e0a3.

+++ src/main/java/frc/robot/commands/liftGotoCommand.java
@@ -11,8 +11,8 @@ import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
 public class liftGotoCommand extends Command {
-  double position;
-  public liftGotoCommand(double p) {
+  int position;
+  public liftGotoCommand(int p) {
     // Use requires() here to declare subsystem dependencies
     // eg. requires(chassis);
      position = p;
